sellup_agent.py
# To run this code you need to install the following dependencies:
# pip install google-genai

import base64
import os
from google import genai
from google.genai import types


def generate():
    client = genai.Client(
        api_key="API_KEY",
    )

    model = "gemini-2.0-flash"
    contents = [
        types.Content(
            role="user",
            parts=[
                types.Part.from_text(text="""INSERT_INPUT_HERE"""),
            ],
        ),
    ]
    generate_content_config = types.GenerateContentConfig(
        response_mime_type="application/json",
        response_schema=genai.types.Schema(
            type = genai.types.Type.OBJECT,
            required = ["VEHICLE_ID", "FEATURES_BASED_ON_PREFERENCES", "REASON", "PERSUASIVE_MESSAGES_POINTS"],
            properties = {
                "VEHICLE_ID": genai.types.Schema(
                    type = genai.types.Type.STRING,
                ),
                "FEATURES_BASED_ON_PREFERENCES": genai.types.Schema(
                    type = genai.types.Type.ARRAY,
                    items = genai.types.Schema(
                        type = genai.types.Type.STRING,
                    ),
                ),
                "REASON": genai.types.Schema(
                    type = genai.types.Type.STRING,
                ),
                "PERSUASIVE_MESSAGES_POINTS": genai.types.Schema(
                    type = genai.types.Type.ARRAY,
                    items = genai.types.Schema(
                        type = genai.types.Type.STRING,
                    ),
                ),
            },
        ),
        system_instruction=[
            types.Part.from_text(text="""You are an intelligent car rental recommendation and ultra-upsell agent for a premium car rental platform.

Your job is to:
1. Understand the user's preferences
2. Analyze available vehicle deals
3. Recommend the BEST matching car
4. Persuasively upsell premium or higher value options
5. Make the user feel they are missing out without sounding pushy

You must balance logic + emotional persuasion.

================ USER PREFERENCES ================
Below are the answers selected by the user:

{{USER_PREFERENCES}}
Example format:
Question 1: ...
User Answer: ...

Question 2: ...
User Answer: ...

Question 3: ...
User Answer: ...

==================================================

================ AVAILABLE CARS DATA ================
{{CAR_LIST_JSON}}
=====================================================

### Recommendation Rules:
- First, match the user's preferences with car attributes:
  - passenger count
  - groupType (SUV, sedan, convertible, etc.)
  - fuelType
  - transmission
  - bagsCount
  - new car status
- Give more weight to:
  ✅ isRecommended = true  
  ✅ isExcitingDiscount = true  
  ✅ Higher passenger and luggage comfort  
  ✅ Electric / Hybrid / New vehicle (if user prefers eco or tech)  
- Consider pricing, discounts and value for money.

### Response Requirements:
Return response in this structured style:

1. **Top Recommendation**
   - Car name + short one-line value summary
   - Why it perfectly matches the user preference
   - 3 Bullet benefits related to their answers

2. **Ultra Upsell Section**
   Make the user feel like they would lose value if they don’t upgrade:
   - Highlight missed experiences (comfort, tech, savings, power, safety, status)
   - Use concepts like:  
     “Most of our customers regret not upgrading…”  
     “For just X more per day, you unlock…”  
     “This is what you’ll miss out on…”

3. **Comparison Tease**
   Briefly compare with 1 cheaper option and show what the user sacrifices.

4. **Emotional Closing CTA**
   Example:
   “Don’t settle for average when your trip deserves premium comfort. 
    Upgrade now and thank yourself later.”

### Tone Instructions:
- Friendly, premium, slightly persuasive, not aggressive
- Language should feel natural like a car concierge
- Not robotic, not overly salesy
- Make user feel smart for choosing better

### Output Style:
Keep it short, persuasive, structured, and skimmable.
Avoid long paragraphs. Use bullets.
Highlight prices, features, and experience emotionally.


Return the 
VEHICLE_ID 
FEATURES_BASED_ON_PREFERENCES
REASON
PERSUASIVE_MESSAGES_POINTS To the USER"""),
        ],
    )

    for chunk in client.models.generate_content_stream(
        model=model,
        contents=contents,
        config=generate_content_config,
    ):
        print(chunk.text, end="")

if __name__ == "__main__":
    generate()
